change unsigned to size_t where approprite
This is the result of questioning int vs unsigned.
Existing code prefers int unless there is a good reason for unsigned.
The usual good reason for unsigned was compatibility with library size_t.
So in those cases code is changed here to actually use size_t.
Otherwise, existing preference of int over unsigned remains.

+++ apps/bm_exp.cc
@@ -186,8 +186,8 @@ TIME_PAIR EVAL_BM_EXP::tr_review(COMPONENT* d)const
 /*--------------------------------------------------------------------------*/
 bool EVAL_BM_EXP::parse_numlist(CS& cmd)
 {
-  unsigned start = cmd.cursor();
-  unsigned here = cmd.cursor();
+  size_t start = cmd.cursor();
+  size_t here = cmd.cursor();
   for (PARAMETER<double>* i = &_iv;  i < &_end;  ++i) {
     PARAMETER<double> val(NOT_VALID);
     cmd >> val;
