+++ apps/measure_at.cc
@@ -35,7 +35,7 @@ public:
     PARAMETER<double> x;
     bool derivative = false;
 
-    unsigned here = Cmd.cursor();
+    size_t here = Cmd.cursor();
     Cmd >> probe_name;
     WAVE* w = find_wave(probe_name);
